+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
   hooks:
     - id: mypy
       args: [--strict, --ignore-missing-imports]
-      exclude: ^(docs/|detectmatelibrary_tests/)
+      exclude: ^(docs/|tests/)
       additional_dependencies:
         - pydantic
         - types-PyYAML
@@ -58,7 +58,7 @@ repos:
   rev: 1.8.6
   hooks:
     - id: bandit
-      exclude: (^detectmatelibrary_tests/|.*/test_.*|^test_.*)
+      exclude: (^tests/|.*/test_.*|^test_.*)
 
 # Unused code detection
 - repo: https://github.com/jendrikseipp/vulture
