+++ docs/parsers/template_matcher.md
@@ -80,7 +80,7 @@ cfg = {
         "MatcherParser": {
             "method_type": "matcher_parser",
             "params": {
-                "path_templates": "detectmatelibrary_tests/test_folder/test_templates.txt",
+                "path_templates": "tests/test_folder/test_templates.txt",
                 "remove_spaces": True,
                 "remove_punctuation": True,
                 "lowercase": True
