+++ tests/test_common/test_config_roundtrip.py
@@ -25,7 +25,7 @@ class MockupDetectorConfig(BasicConfig):
 
 
 def load_test_config() -> dict:
-    with open("detectmatelibrary_tests/test_folder/test_config.yaml", 'r') as file:
+    with open("tests/test_folder/test_config.yaml", 'r') as file:
         return yaml.safe_load(file)
