+++ tests/test_common/test_core.py
