+++ tests/test_common/test_core_detector.py
