+++ tests/test_common/test_extract_timestamp.py
