+++ tests/test_detectors/dummy_detector.py
