+++ tests/test_detectors/test_bigram_frequency_detector.py
@@ -18,7 +18,7 @@ from detectmatelibrary.parsers.template_matcher import MatcherParser
 from detectmatelibrary.helper.from_to import From
 import detectmatelibrary.schemas as schemas
 from detectmatelibrary.utils.aux import time_test_mode
-from detectmatelibrary_tests.test_pipelines.test_configuration_engine import AUDIT_LOG
+from tests.test_pipelines.test_configuration_engine import AUDIT_LOG
 # Set time test mode for consistent timestamps
 time_test_mode()
 
@@ -226,7 +226,7 @@ _PARSER_CONFIG = {
                 "remove_spaces": True,
                 "remove_punctuation": True,
                 "lowercase": True,
-                "path_templates": "detectmatelibrary_tests/test_folder/audit_templates.txt",
+                "path_templates": "tests/test_folder/audit_templates.txt",
             },
         }
     }
