+++ tests/test_detectors/test_new_value_detector.py
@@ -1,22 +1,23 @@
-"""Tests for CharsetDetector class.
+"""Tests for NewValueDetector class.
 
-This module tests the CharsetDetector implementation including:
+This module tests the NewValueDetector implementation including:
 - Initialization and configuration
-- Training functionality to learn known characters
-- Detection logic for new/unknown characters
+- Training functionality to learn known values
+- Detection logic for new/unknown values
 - Event-specific configuration handling
 - Input/output schema validation
 """
 
 from detectmatelibrary.common._core_op._fit_logic import TrainState
-from detectmatelibrary.detectors.charset_detector import CharsetDetector, CharsetDetectorConfig, BufferMode
+from detectmatelibrary.detectors.new_value_detector import NewValueDetector, NewValueDetectorConfig, \
+    BufferMode
 from detectmatelibrary.common._core_op._fit_logic import ConfigState
 from detectmatelibrary.constants import GLOBAL_EVENT_ID
 from detectmatelibrary.parsers.template_matcher import MatcherParser
 from detectmatelibrary.helper.from_to import From
 import detectmatelibrary.schemas as schemas
 from detectmatelibrary.utils.aux import time_test_mode
-from detectmatelibrary_tests.test_pipelines.test_configuration_engine import AUDIT_LOG
+from tests.test_pipelines.test_configuration_engine import AUDIT_LOG
 
 # Set time test mode for consistent timestamps
 time_test_mode()
@@ -25,7 +26,7 @@ time_test_mode()
 config = {
     "detectors": {
         "CustomInit": {
-            "method_type": "charset_detector",
+            "method_type": "new_value_detector",
             "auto_config": False,
             "params": {},
             "events": {
@@ -40,7 +41,7 @@ config = {
             }
         },
         "MultipleDetector": {
-            "method_type": "charset_detector",
+            "method_type": "new_value_detector",
             "auto_config": False,
             "params": {},
             "events": {
@@ -61,14 +62,14 @@ config = {
 }
 
 
-class TestCharsetDetectorInitialization:
-    """Test CharsetDetector initialization and configuration."""
+class TestNewValueDetectorInitialization:
+    """Test NewValueDetector initialization and configuration."""
 
     def test_default_initialization(self):
         """Test detector initialization with default parameters."""
-        detector = CharsetDetector()
+        detector = NewValueDetector()
 
-        assert detector.name == "CharsetDetector"
+        assert detector.name == "NewValueDetector"
         assert hasattr(detector, 'config')
         assert detector.data_buffer.mode == BufferMode.NO_BUF
         assert detector.input_schema == schemas.ParserSchema
@@ -77,44 +78,19 @@ class TestCharsetDetectorInitialization:
 
     def test_custom_config_initialization(self):
         """Test detector initialization with custom configuration."""
-        detector = CharsetDetector(name="CustomInit", config=config)
+        detector = NewValueDetector(name="CustomInit", config=config)
 
         assert detector.name == "CustomInit"
         assert hasattr(detector, 'persistency')
         assert isinstance(detector.persistency.events_data, dict)
 
-    def test_persistency_uses_custom_add_value(self):
-        """Main persistency must accumulate characters; auto_conf must not."""
-        detector = CharsetDetector()
-        # Ingest a sample so a SingleStabilityTracker is materialized
-        detector.persistency.ingest_event(
-            event_id=1,
-            event_template="t",
-            named_variables={"v": "hello"},
-        )
-        single = detector.persistency.get_event_data(1)["v"]
-        assert single.unique_set == {"h", "e", "l", "o"}
 
-    def test_register_persistency_was_called(self):
-        """Main persistency should be registered so persist/load round-trips
-        work."""
-        from detectmatelibrary.common.detector import PersistConfig
-        cfg = CharsetDetectorConfig(
-            persist=PersistConfig(path="memory://charset_regpersist/state")
-        )
-        detector = CharsetDetector(config=cfg)
-        # _register_persistency builds a PersistencySaver bound to detector.persistency
-        assert detector.saver is not None
-        assert detector.saver._persistency is detector.persistency
-        detector.saver.stop()
-
-
-class TestCharsetDetectorTraining:
-    """Test CharsetDetector training functionality."""
+class TestNewValueDetectorTraining:
+    """Test NewValueDetector training functionality."""
 
     def test_train_multiple_values(self):
         """Test training with multiple different values."""
-        detector = CharsetDetector(config=config, name="MultipleDetector")
+        detector = NewValueDetector(config=config, name="MultipleDetector")
         # Train with multiple values (only event 1 should be tracked per config)
         for event in range(3):
             for level in ["INFO", "WARNING", "ERROR"]:
@@ -135,18 +111,19 @@ class TestCharsetDetectorTraining:
         assert len(detector.persistency.events_data) == 1
         event_data = detector.persistency.get_event_data(1)
         assert event_data is not None
-        # With expand_value=True, unique_set contains individual characters
-        assert set("INFO") <= event_data["level"].unique_set
-        assert set("WARNING") <= event_data["level"].unique_set
-        assert set("ERROR") <= event_data["level"].unique_set
-        assert set("assa") <= event_data["test"].unique_set
+        # Check the level values
+        assert "INFO" in event_data["level"].unique_set
+        assert "WARNING" in event_data["level"].unique_set
+        assert "ERROR" in event_data["level"].unique_set
+        # Check the variable at position 1 (named "test")
+        assert "assa" in event_data["test"].unique_set
 
 
-class TestCharsetDetectorDetection:
-    """Test CharsetDetector detection functionality."""
+class TestNewValueDetectorDetection:
+    """Test NewValueDetector detection functionality."""
 
     def test_detect_known_value_no_alert(self):
-        detector = CharsetDetector(config=config, name="MultipleDetector")
+        detector = NewValueDetector(config=config, name="MultipleDetector")
 
         # Train with a value
         train_data = schemas.ParserSchema({
@@ -167,7 +144,7 @@ class TestCharsetDetectorDetection:
             "parserType": "test",
             "EventID": 12,
             "template": "test template",
-            "variables": ["adsasddddddaaa"],
+            "variables": ["adsasd"],
             "logID": "2",
             "parsedLogID": "2",
             "parserID": "test_parser",
@@ -182,7 +159,7 @@ class TestCharsetDetectorDetection:
         assert output.score == 0.0
 
     def test_detect_known_value_alert(self):
-        detector = CharsetDetector(config=config, name="MultipleDetector")
+        detector = NewValueDetector(config=config, name="MultipleDetector")
 
         # Train with a value
         train_data = schemas.ParserSchema({
@@ -203,7 +180,7 @@ class TestCharsetDetectorDetection:
             "parserType": "test",
             "EventID": 1,
             "template": "test template",
-            "variables": ["asas", "adsd"],
+            "variables": ["adsasd", "asdasd"],
             "logID": "2",
             "parsedLogID": "2",
             "parserID": "test_parser",
@@ -217,56 +194,6 @@ class TestCharsetDetectorDetection:
         assert result
         assert output.score == 1.0
 
-    def test_detect_unknown_chars_reported_per_variable(self):
-        """Train on a known alphabet; detect a value with unknown chars and
-        confirm the alert string lists the unknown chars sorted."""
-        cfg = {
-            "detectors": {
-                "Single": {
-                    "method_type": "charset_detector",
-                    "auto_config": False,
-                    "params": {},
-                    "events": {
-                        1: {
-                            "test": {
-                                "params": {},
-                                "variables": [{"pos": 0, "name": "v", "params": {}}],
-                            }
-                        }
-                    },
-                }
-            }
-        }
-        detector = CharsetDetector(config=cfg, name="Single")
-
-        train = schemas.ParserSchema({
-            "parserType": "test", "EventID": 1, "template": "t",
-            "variables": ["abc"], "logID": "1", "parsedLogID": "1",
-            "parserID": "p", "log": "l", "logFormatVariables": {},
-        })
-        detector.train(train)
-
-        # All known chars
-        ok = schemas.ParserSchema({
-            "parserType": "test", "EventID": 1, "template": "t",
-            "variables": ["cba"], "logID": "2", "parsedLogID": "2",
-            "parserID": "p", "log": "l", "logFormatVariables": {},
-        })
-        out = schemas.DetectorSchema()
-        assert not detector.detect(ok, out)
-        assert out.score == 0.0
-
-        # Unknown chars 'x' and 'y'
-        bad = schemas.ParserSchema({
-            "parserType": "test", "EventID": 1, "template": "t",
-            "variables": ["axy"], "logID": "3", "parsedLogID": "3",
-            "parserID": "p", "log": "l", "logFormatVariables": {},
-        })
-        out = schemas.DetectorSchema()
-        assert detector.detect(bad, out)
-        assert out.score == 1.0
-        assert any("'x'" in msg and "'y'" in msg for msg in out["alertsObtain"].values())
-
 
 _PARSER_CONFIG = {
     "parsers": {
@@ -279,19 +206,19 @@ _PARSER_CONFIG = {
                 "remove_spaces": True,
                 "remove_punctuation": True,
                 "lowercase": True,
-                "path_templates": "detectmatelibrary_tests/test_folder/audit_templates.txt",
+                "path_templates": "tests/test_folder/audit_templates.txt",
             },
         }
     }
 }
 
 
-class TestCharsetDetectorEndToEnd:
+class TestNewValueDetectorEndToEnd:
     """Regression test: full configure/train/detect pipeline on audit.log."""
 
     def test_audit_log_anomalies(self):
         parser = MatcherParser(config=_PARSER_CONFIG)
-        detector = CharsetDetector()
+        detector = NewValueDetector()
 
         logs = list(From.log(parser, in_path=AUDIT_LOG, do_process=True))
 
@@ -311,13 +238,13 @@ class TestCharsetDetectorEndToEnd:
         assert detected_ids == {'1859', '1860', '1861', '1862', '1864', '1865', '1866', '1867'}
 
 
-class TestCharsetDetectorAutoConfig:
+class TestNewValueDetectorAutoConfig:
     """Test that process() drives configure/set_configuration/train/detect
     automatically."""
 
     def test_audit_log_anomalies_via_process(self):
         parser = MatcherParser(config=_PARSER_CONFIG)
-        detector = CharsetDetector()
+        detector = NewValueDetector()
 
         logs = list(From.log(parser, in_path=AUDIT_LOG, do_process=True))
 
@@ -344,7 +271,7 @@ class TestCharsetDetectorAutoConfig:
         assert detected_ids == {'1859', '1860', '1861', '1862', '1864', '1865', '1866', '1867'}
 
 
-class TestCharsetDetectorGlobalInstances:
+class TestNewValueDetectorGlobalInstances:
     """Tests event-ID-independent global instance detection."""
 
     def test_global_instance_detects_new_type(self):
@@ -353,10 +280,9 @@ class TestCharsetDetectorGlobalInstances:
         parser = MatcherParser(config=_PARSER_CONFIG)
         config_dict = {
             "detectors": {
-                "CharsetDetector": {
-                    "method_type": "charset_detector",
+                "NewValueDetector": {
+                    "method_type": "new_value_detector",
                     "auto_config": False,
-                    "params": {},
                     "global": {
                         "test": {
                             "header_variables": [{"pos": "Type"}]
@@ -365,8 +291,8 @@ class TestCharsetDetectorGlobalInstances:
                 }
             }
         }
-        config = CharsetDetectorConfig.from_dict(config_dict, "CharsetDetector")
-        detector = CharsetDetector(config=config)
+        config = NewValueDetectorConfig.from_dict(config_dict, "NewValueDetector")
+        detector = NewValueDetector(config=config)
 
         logs = list(From.log(parser, in_path=AUDIT_LOG, do_process=True))
 
@@ -384,27 +310,3 @@ class TestCharsetDetectorGlobalInstances:
                 detected_ids.add(log["logID"])
 
         assert len(detected_ids) > 0
-
-
-class TestCharsetDetectorSetConfigurationPreservesPersist:
-    def test_persist_flag_survives_set_configuration(self):
-        from detectmatelibrary.common.detector import PersistConfig
-
-        detector = CharsetDetector()
-        # Simulate persist being enabled by an earlier config load
-        detector.config.persist = PersistConfig(path="./state")
-
-        # Feed configure() with a couple of stable-variable samples
-        for _ in range(5):
-            sample = schemas.ParserSchema({
-                "parserType": "test", "EventID": 1, "template": "t",
-                "variables": ["abc"], "logID": "x", "parsedLogID": "x",
-                "parserID": "p", "log": "l",
-                "logFormatVariables": {"level": "INFO"},
-            })
-            detector.configure(sample)
-
-        detector.set_configuration()
-
-        assert detector.config.persist is not None
-        assert detector.config.persist.path == "./state"
