+++ tests/test_detectors/test_persist_integration.py
