+++ tests/test_detectors/test_random_detector.py
