+++ tests/test_detectors/test_rule_detector.py
