+++ tests/test_detectors/test_value_range_detector.py
@@ -20,7 +20,7 @@ from detectmatelibrary.helper.from_to import From
 import detectmatelibrary.schemas as schemas
 from detectmatelibrary.utils.aux import time_test_mode
 from detectmatelibrary_tools.logging import logger
-from detectmatelibrary_tests.test_pipelines.test_configuration_engine import AUDIT_LOG
+from tests.test_pipelines.test_configuration_engine import AUDIT_LOG
 # Set time test mode for consistent timestamps
 time_test_mode()
 
@@ -315,7 +315,7 @@ _PARSER_CONFIG = {
                 "remove_spaces": True,
                 "remove_punctuation": True,
                 "lowercase": True,
-                "path_templates": "detectmatelibrary_tests/test_folder/audit_templates.txt",
+                "path_templates": "tests/test_folder/audit_templates.txt",
             },
         }
     }
