+++ tests/test_persistency/test_stability_tracker.py
