+++ tests/test_pipelines/test_configuration_engine.py
@@ -4,9 +4,9 @@ from detectmatelibrary.helper.from_to import From
 
 import json
 
-AUDIT_LOG = "detectmatelibrary_tests/test_folder/audit.log"
-AUDIT_TEMPLATES = "detectmatelibrary_tests/test_folder/audit_templates.txt"
-ANOMALY_LABELS = "detectmatelibrary_tests/test_folder/audit_anomaly_labels.log"
+AUDIT_LOG = "tests/test_folder/audit.log"
+AUDIT_TEMPLATES = "tests/test_folder/audit_templates.txt"
+ANOMALY_LABELS = "tests/test_folder/audit_anomaly_labels.log"
 LOG_FORMAT = "type=<Type> msg=audit(<Time>): <Content>"
 TRAIN_UNTIL = 1800
