+++ tests/test_pipelines/test_dummy_methods.py
@@ -1,6 +1,6 @@
 
-from detectmatelibrary_tests.test_detectors.dummy_detector import DummyDetector, DummyDetectorConfig
-from detectmatelibrary_tests.test_parsers.dummy_parser import DummyParser, DummyParserConfig
+from tests.test_detectors.dummy_detector import DummyDetector, DummyDetectorConfig
+from tests.test_parsers.dummy_parser import DummyParser, DummyParserConfig
 import detectmatelibrary.schemas as schemas
