+++ tests/test_schemas/test_schema_class.py
