+++ tests/test_utils/test_stability_tracking.py
